requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
langchain>=0.1.0
langchain-huggingface>=0.0.1

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
ta>=0.10.2

# Embeddings and Search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# API clients
requests>=2.31.0
finnhub-python>=2.4.0

# Utilities
python-dotenv>=1.0.0